migrate styled components into styles dir to resemble stylesheet
