+++ src/App/Routes.jsx
@@ -8,7 +8,7 @@ const Routes = () => (
   <Switch>
     <Route exact path="/login" component={Screens.Login} />
     <Route exact path="/signup" component={Screens.Login} />
-    <Route exact path="/learn" component={Screens.LearnMore} />
+    <Route exact path="/learn" component={Screens.Learn} />
     <Route exact path="/locations" component={Screens.Locations} />
     <Route exact path="/users/:username" component={Screens.Users} />
     <AuthRoute exact path="/groups" component={Screens.Groups} />
