+++ src/components/atoms/AngleBracket/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const AngleBracket = () => <span>&gt;</span>;
+
+export default AngleBracket;
