--- src/components/atoms/Input/index.jsx
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-const Input = styled.input``;
-
-export default Input;
