--- src/components/atoms/Label/index.jsx
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-const Label = styled.label``;
-
-export default Label;
