--- src/components/atoms/Paragraph/index.jsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-const Paragraph = styled.p`
-  font-size: 1.2em;
-  line-height: 1.4;
-  margin: 20px 1%;
-`;
-
-export default Paragraph;
