--- src/components/atoms/StyledLink/StyledLink.test.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-import MyComponent from "../StyledLink";
-
-const ExampleComponent = () => <div />;
-
-it("renders without crashing", () => {
-  shallow(
-    <MyComponent to="/">
-      <ExampleComponent />
-    </MyComponent>
-  );
-});
