+++ src/components/atoms/index.js
@@ -8,9 +8,7 @@ export { default as IconNav } from "./IconNav";
 export { default as IconSub } from "./IconSub";
 export { default as Image } from "./Image";
 export { default as ImagePlaceholder } from "./ImagePlaceholder";
-export { default as Input } from "./Input";
-export { default as Label } from "./Label";
 export { default as ListLink } from "./ListLink";
-export { default as Paragraph } from "./Paragraph";
 export { default as StyledLink } from "./StyledLink";
 export { default as TextVisibility } from "./TextVisibility";
+export { default as AngleBracket } from "./AngleBracket";
