+++ src/components/molecules/index.js
@@ -1,3 +1,3 @@
 export { default as Comment } from "./Comment";
-export { default as DevLoginButton } from "./DevLoginButton";
+export { default as LoginBtn } from "./LoginBtn";
 export { default as Post } from "./Post";
