+++ src/components/screens/Landing/NewestItem.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import { Paragraph, StyledLink } from "components/atoms";
+import { AngleBracket } from "components/atoms";
+import { StyledLink, Paragraph, StyledParagraph } from "styles";
 
 const NewestItem = () => (
   <StyledParagraph>
@@ -10,15 +11,5 @@ const NewestItem = () => (
     </StyledLink>
   </StyledParagraph>
 );
-
-const AngleBracket = () => <span>&gt;</span>;
-
-const StyledParagraph = styled(Paragraph)`
-  font-size: 0.8em;
-  background-color: ${({ theme }) => theme.colors.primaryDark};
-  margin: 4% 3% auto;
-  padding: 3px 10px;
-  border-radius: 20px;
-`;
 
 export default NewestItem;
