+++ src/components/screens/Landing/index.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
+import { StyledContainer, Panel } from "styles";
+
 import CallToActions from "./CallToActions";
 import HeroText from "./HeroText";
 import NewestItem from "./NewestItem";
@@ -14,38 +16,8 @@ const Landing = () => (
       <Tagline />
       <CallToActions />
     </Panel>
-    <Panel isWhite>
-      <p style={{ margin: 30 }}>Image</p>
-    </Panel>
+    <Panel isWhite>{/* image placeholder */}</Panel>
   </StyledContainer>
 );
-
-const StyledContainer = styled.main`
-  /* display */
-  display: flex;
-  flex-wrap: wrap;
-  height: 100%;
-  flex: 1;
-
-  /* font */
-  font-size: 1.5em;
-  line-height: 1.4;
-  font-family: ${({ theme }) => theme.fonts.secondary};
-`;
-
-const Panel = styled.div`
-  /* display */
-  flex: 1;
-
-  /* box */
-  min-width: 425px;
-  padding-top: 3%;
-
-  /* color */
-  color: ${({ isWhite, theme }) =>
-    isWhite ? theme.colors.primary : theme.colors.white};
-  background-color: ${({ isWhite, theme }) =>
-    isWhite ? theme.colors.white : theme.colors.primary};
-`;
 
 export default Landing;
