+++ src/components/screens/Learn/Learn.stories.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { storiesOf } from "@storybook/react";
+
+import MyComponent from "../Learn";
+
+storiesOf("Screens", module).add("Learn", () => <MyComponent />);
