+++ src/components/screens/Learn/Learn.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import MyComponent from "../LearnMore";
+import MyComponent from "../Learn";
 
 it("renders without crashing", () => {
   shallow(<MyComponent />);
