--- src/components/screens/LearnMore/LearnMore.stories.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import { storiesOf } from "@storybook/react";
-
-import MyComponent from "../LearnMore";
-
-storiesOf("Screens", module).add("LearnMore", () => <MyComponent />);
